+++ pf1.js
@@ -189,15 +189,60 @@ Hooks.once("setup", function () {
     "actorStatures",
   ];
 
-  const doLocalize = function (obj) {
-    return Object.entries(obj).reduce((obj, e) => {
-      if (typeof e[1] === "string") obj[e[0]] = game.i18n.localize(e[1]);
-      else if (typeof e[1] === "object") obj[e[0]] = doLocalize(e[1]);
+  // Config (sub-)objects to be sorted
+  const toSort = [
+    "buffTargets",
+    "buffTargets.misc",
+    "contextNoteTargets",
+    "contextNoteTargets.misc",
+    "skills",
+    "conditions",
+    "conditionTypes",
+    "consumableTypes",
+    "creatureTypes",
+    "featTypes",
+    "weaponProperties",
+    "spellSchools",
+    "languages",
+    "damageTypes",
+  ];
+
+  /**
+   * Helper function to recursively localize object entries
+   *
+   * @param {object} obj - The object to be localized
+   * @param {string} cat - The object's name
+   * @returns {object} The localized object
+   */
+  const doLocalize = (obj, cat) => {
+    // Create tuples of (key, localized object/string)
+    const localized = Object.entries(obj).reduce((arr, e) => {
+      if (typeof e[1] === "string") arr.push([e[0], game.i18n.localize(e[1])]);
+      else if (typeof e[1] === "object") arr.push([e[0], doLocalize(e[1], `${cat}.${e[0]}`)]);
+      return arr;
+    }, []);
+    if (toSort.includes(cat)) {
+      // Sort simple strings, fall back to sorting by label for objects/categories
+      localized.sort((a, b) => {
+        const localA = typeof a?.[1] === "string" ? a[1] : a[1]?._label;
+        const localB = typeof b?.[1] === "string" ? b[1] : b[1]?._label;
+        // Move misc to bottom of every list
+        if (a[0] === "misc") return 1;
+        else if (b[0] === "misc") return -1;
+        // Regular sorting of localized strings
+        return localA.localeCompare(localB);
+      });
+    }
+    // Get the localized and sorted object out of tuple
+    return localized.reduce((obj, e) => {
+      obj[e[0]] = e[1];
       return obj;
     }, {});
   };
-  for (let o of toLocalize) {
-    CONFIG.PF1[o] = doLocalize(CONFIG.PF1[o]);
+
+  // Localize and sort CONFIG objects
+  for (const o of toLocalize) {
+    CONFIG.PF1[o] = doLocalize(CONFIG.PF1[o], o);
   }
 
   // TinyMCE variables and commands